+++ examples/page_trait/src/page.rs
@@ -1,0 +1,3 @@
+mod create_page;
+
+crate::create_page![my_page::MyPage, my_page_2::MyPage2];